requirements.txt
streamlit
streamlit-searchbox
numpy
scipy
trimesh>=4.0.0
Pillow
requests
pyvista
scikit-image
pathlib
libgl1-mesa-glx
libglu1-mesa
xvfb
libxrender1
libxext6
 

